# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DentSalud – Sistema de Gestión Odontológica API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pacientes/{id}/plan-tratamientos:
    get:
      operationId: getPlanTratamientos
      tags: [tratamientos]
      summary: Get treatment plan for patient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Treatment plan items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanTratamiento"
    post:
      operationId: createPlanTratamiento
      tags: [tratamientos]
      summary: Add treatment to plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanTratamientoInput"
      responses:
        "201":
          description: Created plan item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanTratamiento"
  /plan-tratamientos/{id}:
    patch:
      operationId: updatePlanTratamiento
      tags: [tratamientos]
      summary: Update treatment plan item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanTratamientoUpdate"
      responses:
        "200":
          description: Updated plan item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanTratamiento"
    delete:
      operationId: deletePlanTratamiento
      tags: [tratamientos]
      summary: Delete treatment plan item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /tipos-tratamiento:
    get:
      operationId: listTiposTratamiento
      tags: [tratamientos]
      summary: List all treatment types
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Treatment types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TipoTratamiento"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TipoTratamiento:
      type: object
      required: [id, nombre, precioReferencia]
      properties:
        id:
          type: integer
        nombre:
          type: string
        subtipo:
          type: ["string", "null"]
        codigo:
          type: ["string", "null"]
        precioReferencia:
          type: number
        descripcion:
          type: ["string", "null"]
        activo:
          type: boolean
    PlanTratamiento:
      type: object
      required: [id, pacienteId, tipoTratamientoId, cantidad, precioUnitario, estado]
      properties:
        id:
          type: integer
        pacienteId:
          type: integer
        tipoTratamientoId:
          type: integer
        tipoTratamiento:
          $ref: "#/components/schemas/TipoTratamiento"
        numeroPieza:
          type: ["string", "null"]
        subtipoDetalle:
          type: ["string", "null"]
        cantidad:
          type: integer
        precioUnitario:
          type: number
        descuento:
          type: number
        total:
          type: number
        estado:
          type: string
          enum: [pendiente, en_proceso, completado, cancelado]
        prioridad:
          type: ["integer", "null"]
        notas:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PlanTratamientoInput:
      type: object
      required: [tipoTratamientoId, cantidad, precioUnitario]
      properties:
        tipoTratamientoId:
          type: integer
        numeroPieza:
          type: ["string", "null"]
        subtipoDetalle:
          type: ["string", "null"]
        cantidad:
          type: integer
        precioUnitario:
          type: number
        descuento:
          type: number
        estado:
          type: string
          enum: [pendiente, en_proceso, completado, cancelado]
        prioridad:
          type: ["integer", "null"]
        notas:
          type: ["string", "null"]
    PlanTratamientoUpdate:
      type: object
      properties:
        cantidad:
          type: integer
        precioUnitario:
          type: number
        descuento:
          type: number
        estado:
          type: string
          enum: [pendiente, en_proceso, completado, cancelado]
        prioridad:
          type: ["integer", "null"]
        notas:
          type: ["string", "null"]
        numeroPieza:
          type: ["string", "null"]
        subtipoDetalle:
          type: ["string", "null"]
...