# lib/api-zod/src/generated/types/planTratamientoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DentSalud – Sistema de Gestión Odontológica API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanTratamientoInputEstado } from './planTratamientoInputEstado';

export interface PlanTratamientoInput {
  tipoTratamientoId: number;
  /** @nullable */
  numeroPieza?: string | null;
  /** @nullable */
  subtipoDetalle?: string | null;
  cantidad: number;
  precioUnitario: number;
  descuento?: number;
  estado?: PlanTratamientoInputEstado;
  /** @nullable */
  prioridad?: number | null;
  /** @nullable */
  notas?: string | null;
}