# lib/api-zod/src/generated/types/planTratamientoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DentSalud – Sistema de Gestión Odontológica API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanTratamientoUpdateEstado } from './planTratamientoUpdateEstado';

export interface PlanTratamientoUpdate {
  cantidad?: number;
  precioUnitario?: number;
  descuento?: number;
  estado?: PlanTratamientoUpdateEstado;
  /** @nullable */
  prioridad?: number | null;
  /** @nullable */
  notas?: string | null;
  /** @nullable */
  numeroPieza?: string | null;
  /** @nullable */
  subtipoDetalle?: string | null;
}